distribution: Refactor test_distribution_version testcases
* Read expected data from json files rather than editing testcase
