+++ changelogs/fragments/distribution_test_refactor.yml
@@ -1,0 +1,2 @@
+minor_changes:
+- Refactor test_distribution_version testcases.
