+++ test/units/module_utils/facts/system/distribution/fixtures/centos_6.7.json
@@ -1,0 +1,31 @@
+{
+    "name": "CentOS 6.7",
+    "platform.dist": ["centos", "6.7", "Final"],
+    "distro": {
+        "codename": "Final",
+        "id": "centos",
+        "name": "CentOS Linux",
+        "version": "6.7",
+        "version_best": "6.7",
+        "os_release_info": {},
+        "lsb_release_info": {
+            "release": "6.7",
+            "codename": "Final",
+            "distributor_id": "CentOS",
+            "lsb_version": ":base-4.0-amd64:base-4.0-noarch:core-4.0-amd64:core-4.0-noarch",
+            "description": "CentOS release 6.7 (Final)"
+        }
+    },
+    "input": {
+        "/etc/redhat-release": "CentOS release 6.7 (Final)\n",
+        "/etc/lsb-release": "LSB_VERSION=base-4.0-amd64:base-4.0-noarch:core-4.0-amd64:core-4.0-noarch:graphics-4.0-amd64:graphics-4.0-noarch:printing-4.0-amd64:printing-4.0-noarch\n",
+        "/etc/system-release": "CentOS release 6.7 (Final)\n"
+    },
+    "result": {
+        "distribution_release": "Final",
+        "distribution": "CentOS",
+        "distribution_major_version": "6",
+        "os_family": "RedHat",
+        "distribution_version": "6.7"
+    }
+}
