+++ test/units/module_utils/facts/system/distribution/fixtures/fedora_22.json
@@ -1,0 +1,25 @@
+{
+    "name": "Fedora 22",
+    "platform.dist": ["fedora", "22", "Twenty Two"],
+    "distro": {
+        "codename": "Twenty Two",
+        "id": "fedora",
+        "name": "Fedora",
+        "version": "22",
+        "version_best": "22",
+        "os_release_info": {},
+        "lsb_release_info": {}
+    },
+    "input": {
+        "/etc/redhat-release": "Fedora release 22 (Twenty Two)\n",
+        "/etc/os-release": "NAME=Fedora\nVERSION=\"22 (Twenty Two)\"\nID=fedora\nVERSION_ID=22\nPRETTY_NAME=\"Fedora 22 (Twenty Two)\"\nANSI_COLOR=\"0;34\"\nCPE_NAME=\"cpe:/o:fedoraproject:fedora:22\"\nHOME_URL=\"https://fedoraproject.org/\"\nBUG_REPORT_URL=\"https://bugzilla.redhat.com/\"\nREDHAT_BUGZILLA_PRODUCT=\"Fedora\"\nREDHAT_BUGZILLA_PRODUCT_VERSION=22\nREDHAT_SUPPORT_PRODUCT=\"Fedora\"\nREDHAT_SUPPORT_PRODUCT_VERSION=22\nPRIVACY_POLICY_URL=https://fedoraproject.org/wiki/Legal:PrivacyPolicy\n",
+        "/etc/system-release": "Fedora release 22 (Twenty Two)\n"
+    },
+    "result": {
+        "distribution_release": "Twenty Two",
+        "distribution": "Fedora",
+        "distribution_major_version": "22",
+        "os_family": "RedHat",
+        "distribution_version": "22"
+    }
+}
