+++ test/units/module_utils/facts/system/distribution/fixtures/kde_neon_16.04.json
@@ -1,0 +1,42 @@
+{
+    "platform.dist": ["neon", "16.04", "xenial"],
+    "distro": {
+        "codename": "xenial",
+        "id": "neon",
+        "name": "KDE neon",
+        "version": "16.04",
+        "version_best": "16.04",
+        "os_release_info": {
+            "support_url": "http://help.ubuntu.com/",
+            "version_codename": "xenial",
+            "pretty_name": "Ubuntu 16.04.6 LTS",
+            "home_url": "http://www.ubuntu.com/",
+            "bug_report_url": "http://bugs.launchpad.net/ubuntu/",
+            "version": "16.04.6 LTS (Xenial Xerus)",
+            "version_id": "16.04",
+            "id": "ubuntu",
+            "ubuntu_codename": "xenial",
+            "codename": "xenial",
+            "name": "Ubuntu",
+            "id_like": "debian"
+        },
+        "lsb_release_info": {
+            "description": "Ubuntu 16.04.6 LTS",
+            "release": "16.04",
+            "distributor_id": "Ubuntu",
+            "codename": "xenial"
+        }
+    },
+    "input": {
+        "/etc/os-release": "NAME=\"KDE neon\"\nVERSION=\"5.8\"\nID=neon\nID_LIKE=\"ubuntu debian\"\nPRETTY_NAME=\"KDE neon User Edition 5.8\"\nVERSION_ID=\"16.04\"\nHOME_URL=\"http://neon.kde.org/\"\nSUPPORT_URL=\"http://neon.kde.org/\"\nBUG_REPORT_URL=\"http://bugs.kde.org/\"\nVERSION_CODENAME=xenial\nUBUNTU_CODENAME=xenial\n",
+        "/etc/lsb-release": "DISTRIB_ID=neon\nDISTRIB_RELEASE=16.04\nDISTRIB_CODENAME=xenial\nDISTRIB_DESCRIPTION=\"KDE neon User Edition 5.8\"\n"
+    },
+    "name": "KDE neon 16.04",
+    "result": {
+        "distribution_release": "xenial",
+        "distribution": "KDE neon",
+        "distribution_major_version": "16",
+        "os_family": "Debian",
+        "distribution_version": "16.04"
+    }
+}
