+++ test/units/module_utils/facts/system/distribution/fixtures/linux_mint_18.2.json
@@ -1,0 +1,25 @@
+{
+    "platform.dist": ["linuxmint", "18.2", "sonya"],
+    "input": {
+        "/etc/os-release": "NAME=\"Linux Mint\"\nVERSION=\"18.2 (Sonya)\"\nID=linuxmint\nID_LIKE=ubuntu\nPRETTY_NAME=\"Linux Mint 18.2\"\nVERSION_ID=\"18.2\"\nHOME_URL=\"http://www.linuxmint.com/\"\nSUPPORT_URL=\"http://forums.linuxmint.com/\"\nBUG_REPORT_URL=\"http://bugs.launchpad.net/linuxmint/\"\nVERSION_CODENAME=sonya\nUBUNTU_CODENAME=xenial\n",
+        "/usr/lib/os-release": "NAME=\"Linux Mint\"\nVERSION=\"18.2 (Sonya)\"\nID=linuxmint\nID_LIKE=ubuntu\nPRETTY_NAME=\"Linux Mint 18.2\"\nVERSION_ID=\"18.2\"\nHOME_URL=\"http://www.linuxmint.com/\"\nSUPPORT_URL=\"http://forums.linuxmint.com/\"\nBUG_REPORT_URL=\"http://bugs.launchpad.net/linuxmint/\"\nVERSION_CODENAME=sonya\nUBUNTU_CODENAME=xenial\n",
+        "/etc/lsb-release": "DISTRIB_ID=LinuxMint\nDISTRIB_RELEASE=18.2\nDISTRIB_CODENAME=sonya\nDISTRIB_DESCRIPTION=\"Linux Mint 18.2 Sonya\"\n"
+    },
+    "result": {
+        "distribution_release": "sonya",
+        "distribution": "Linux Mint",
+        "distribution_major_version": "18",
+        "os_family": "Debian",
+        "distribution_version": "18.2"
+    },
+    "name": "Linux Mint 18.2",
+    "distro": {
+        "codename": "sonya",
+        "version": "18.2",
+        "id": "linuxmint",
+        "version_best": "18.2",
+        "name": "Linux Mint",
+        "os_release_info": {},
+        "lsb_release_info": {}
+    }
+}
