+++ test/units/module_utils/facts/system/distribution/fixtures/smgl_na.json
@@ -1,0 +1,23 @@
+{
+    "platform.dist": ["", "", ""],
+    "distro": {
+        "codename": "",
+        "id": "smgl",
+        "name": "Source Mage GNU/Linux",
+        "version": "",
+        "version_best": "",
+        "os_release_info": {},
+        "lsb_release_info": {}
+    },
+    "input": {
+        "/etc/sourcemage-release": "Source Mage GNU/Linux x86_64-pc-linux-gnu\nInstalled from tarball using chroot image (Grimoire 0.61-rc) on Thu May 17 17:31:37 UTC 2012\n"
+    },
+    "name": "SMGL NA",
+    "result": {
+        "distribution_release": "NA",
+        "distribution": "SMGL",
+        "distribution_major_version": "NA",
+        "os_family": "SMGL",
+        "distribution_version": "NA"
+    }
+}
