+++ test/units/module_utils/facts/system/distribution/fixtures/solaris_11.3.json
@@ -1,0 +1,25 @@
+{
+    "name": "Solaris 11.3",
+    "uname_r": "5.11",
+    "platform.dist": ["", "", ""],
+    "distro": {
+        "codename": "",
+        "id": "",
+        "name": "",
+        "version": "",
+        "version_best": "",
+        "os_release_info": {},
+        "lsb_release_info": {}
+    },
+    "input": {
+        "/etc/release": "                             Oracle Solaris 11.3 X86\n  Copyright (c) 1983, 2018, Oracle and/or its affiliates.  All rights reserved.\n                              Assembled 09 May 2018\n"
+    },
+    "platform.system": "SunOS",
+    "result": {
+        "distribution_release": "Oracle Solaris 11.3 X86",
+        "distribution": "Solaris",
+        "os_family": "Solaris",
+        "distribution_major_version": "11",
+        "distribution_version": "11.3"
+    }
+}
